ref: Moved status to a constant file
