+++ packages/account/src/Constants/status-codes.js
@@ -1,0 +1,7 @@
+export const income_status_codes = {
+    none: 'none',
+    pending: 'pending',
+    rejected: 'rejected',
+    verified: 'verified',
+    locked: 'locked',
+};
