+++ packages/account/src/Sections/Verification/ProofOfIncome/proof-of-income-container.jsx
@@ -1,6 +1,5 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import { income_status_codes, poinc_documents_list } from './proof-of-income-utils';
 import { Loading, useStateCallback } from '@deriv/components';
 import { WS } from '@deriv/shared';
 import PoincUnverified from 'Components/poinc/status/unverified';
@@ -8,7 +7,9 @@ import PoincReceived from 'Components/poinc/status/received';
 import PoincLimited from 'Components/poinc/status/limited';
 import PoincVerified from 'Components/poinc/status/verified';
 import PoincNotRequired from 'Components/poinc/status/not-required';
+import { income_status_codes } from 'Constants/status-codes';
 import ProofOfIncomeForm from './proof-of-income-form.jsx';
+import { poinc_documents_list } from './proof-of-income-utils';
 import { populateVerificationStatus } from '../Helpers/verification';
 
 const ProofOfIncomeContainer = ({ is_switching, refreshNotifications }) => {
